pyproject.toml
[project]
name = "construct"
version = "0.1.0"
description = "A unit testing framework for VLMs controlling robotic systems"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "odyssey",
    "numpy",
    "google-genai",
    "pillow",
    "python-dotenv",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-asyncio>=0.24"]
dev = ["construct[test]", "ruff", "mypy"]

[project.scripts]
construct = "construct.cli:main"

[project.entry-points.pytest11]
construct = "construct.pytest_plugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
odyssey = { git = "https://github.com/odysseyml/odyssey-python.git" }

[tool.pytest.ini_options]
asyncio_mode = "auto"
